fixed comments user for profile

+++ database/migrations/2014_10_12_000000_create_users_table.php
@@ -12,14 +12,14 @@ return new class extends Migration
     public function up(): void
     {
         Schema::create('users', function (Blueprint $table) {
-            $table->increments('id');
+            $table->bigIncrements('id');
             $table->string('name');
             $table->string('phone')->unique()->default(null);
             $table->string('email')->unique();
             $table->timestamp('email_verified_at')->nullable();
             $table->string('password');
             $table->string('address')->nullable();
-            $table->bigInteger('family_id');
+            $table->unsignedBigInteger('family_id');
             $table->tinyInteger('type')->comment('1: Admin, 2: Teacher, 3: Student');
             $table->tinyInteger('gender')->comment('1: Male, 2: Female');
             $table->string('avatar')->nullable();
